Enable passing matchers to ParamterizedLogging
Part of [#49]
